pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verba-client"
version = "0.1.0"
description = "Python client for the Verba (VFC Capture) HTTP Business API"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "responses>=0.23",
    "python-dotenv>=1.0",
]
